refactor(timeline): Move finding retry indices into DecryptionRetryTask
